backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
requests
google-generativeai
python-dotenv
